requirements.txt
openai
requests
python-dotenv
beautifulsoup4
pandas
numpy
matplotlib
seaborn
scipy
